ekf2: move ev control to file
