+++ src/modules/ekf2/CMakeLists.txt
@@ -57,6 +57,7 @@ px4_add_module(
 		EKF/ekf_helper.cpp
 		EKF/EKFGSF_yaw.cpp
 		EKF/estimator_interface.cpp
+		EKF/ev_height_control.cpp
 		EKF/fake_pos_control.cpp
 		EKF/gps_checks.cpp
 		EKF/gps_control.cpp
